requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
dj-database-url==3.0.1
gunicorn==23.0.0
openpyxl==3.1.5
whitenoise==6.9.0
